perf(ux): split AgentFlowVisualizer into server/client, drop shared JS; remove stray 'use client' in nav; modularize imports; tighten JS budget; centralize no-prefetch nav
