+++ lib/agents/dataAdapter.ts
@@ -1,0 +1,1 @@
+export interface Adapter {}
