+++ package.json
@@ -58,7 +58,7 @@
     "ci-install": "npm ci --prefer-offline --no-audit --no-fund",
     "guard:image": "ts-node --project tsconfig.node.json scripts/guards/imageResponseImportGuard.ts",
     "guard:segments": "ts-node --project tsconfig.node.json scripts/guards/segmentConfigGuard.ts",
-    "analyze": "cross-env ANALYZE=true NODE_ENV=production SPORTS_API_KEY=DUMMY_ANALYZE NEXT_PUBLIC_SUPABASE_URL=http://localhost:54321 NEXT_PUBLIC_SUPABASE_ANON_KEY=dummy-anon next build && node scripts/perf/printBundleHints.mjs",
+    "analyze": "ANALYZE=1 next build",
     "analyze:view": "ANALYZE=true next build && npx http-server .next/analyze -p 5050 -o",
     "perf:report": "node scripts/perf/printBundleHints.mjs",
     "analyze:profile": "cross-env ANALYZE=true NODE_ENV=production ENV_PROFILE=analyze SPORTS_API_KEY=DUMMY_ANALYZE NEXT_PUBLIC_SUPABASE_URL=http://localhost:54321 NEXT_PUBLIC_SUPABASE_ANON_KEY=dummy-anon next build && node scripts/perf/printBundleHints.mjs",
