+++ scripts/guards/jsBudgetGuard.js
@@ -1,16 +1,31 @@
-// Parses Next build output from env var BUILD_SUMMARY (set in npm script)
 const summary = process.env.BUILD_SUMMARY || '';
-const m = /First Load JS shared by all\s+([\d.]+)\s*kB/i.exec(summary);
-const kb = m ? parseFloat(m[1]) : NaN;
-const LIMIT = 500; // adjust as needed
+const pickKb = (label) => {
+  const re = new RegExp(label.replace(/[.*+?^${}()|[\]\\]/g, '\\$&') + '\\s+([\\d.]+)\\s*kB', 'i');
+  const m = re.exec(summary);
+  return m ? parseFloat(m[1]) : NaN;
+};
 
-if (!isNaN(kb)) {
-  if (kb > LIMIT) {
-    console.error(`❌ First Load JS shared by all = ${kb} kB (limit ${LIMIT} kB).`);
-    process.exit(1);
-  } else {
-    console.log(`✅ JS budget ok: ${kb} kB (limit ${LIMIT} kB).`);
+const appKb = pickKb('Route (app)[\\s\\S]*?First Load JS shared by all');
+const pagesKb = pickKb('Route (pages)[\\s\\S]*?First Load JS shared by all');
+const LIMIT_APP = 390;   // step down target
+const LIMIT_PAGES = 430; // step down target
+
+const checks = [
+  ['app', appKb, LIMIT_APP],
+  ['pages', pagesKb, LIMIT_PAGES],
+];
+
+let fail = false;
+for (const [name, kb, limit] of checks) {
+  if (isNaN(kb)) {
+    console.log(`ℹ️ Could not parse "${name}" first-load size; skipping.`);
+    continue;
   }
-} else {
-  console.log('ℹ️ Could not parse first-load size; skipping budget check.');
+  if (kb > limit) {
+    console.error(`❌ ${name} First Load JS = ${kb} kB (limit ${limit} kB).`);
+    fail = true;
+  } else {
+    console.log(`✅ ${name} JS budget ok: ${kb} kB (limit ${limit} kB).`);
+  }
 }
+if (fail) process.exit(1);
